# application/src/main/resources/static/openapi/canyoningApp-openapi.yml
openapi: 3.0.3
info:
  title: Swagger CanyoningApp - OpenAPI 3.0
  description: Swagger of canyoning app
  contact:
    email: delgadofernandez.luis@gmail.com
  version: 0.0.1
servers:
  - url: http://localhost:8080
  #- url: https://misbarrancosbdesarrollo.com:8443
...
paths:
  /createUser:
    post:
      tags:
        - User
      summary: Create a new user
      description: Create a new user
      operationId: createUser
      requestBody:
        description: data from the user
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      security: [ ]
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
        400:
          $ref: '#/components/responses/400BadRequest'
        409:
          $ref: '#/components/responses/409Exists'
  /validateUser:
    get:
      tags:
        - User
      summary: Validate user
      description: Validate an user created withour verify
      operationId: validateUser
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        200:
          description: Successful operation
          content:
            text/plain:
              schema:
                type: string
  /users:
    get:
      tags:
        - User
      summary: Get all users
      description: Get all users filtering bay differents fields
      operationId: getUsers
      parameters:
        - name: email
          in: query
          required: false
          schema:
            type: string
        - name: name
          in: query
          required: false
          schema:
            type: string
        - name: location
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          description: Número de página (empezando desde 0)
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Número de elementos por página
          required: false
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          description: Criterios de ordenación en el formato propiedad(,asc|desc). El orden por defecto es ascendente.
          required: false
          schema:
            type: string
            default: userId,asc
      x-spring-paginated: true
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SimpleUser'
          headers:
            Link:
              $ref: '#/components/headers/Link'
            X-Page-Number:
              $ref: '#/components/headers/X-Page-Number'
            X-Total-Elements:
              $ref: '#/components/headers/X-Total-Elements'
            X-Total-Pages:
              $ref: '#/components/headers/X-Total-Pages'
        400:
          $ref: '#/components/responses/400BadRequest'
  /users/{email}:
    get:
      tags:
        - User
      summary: Get an user
      description: Get user by id
      operationId: getUser
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
        400:
          $ref: '#/components/responses/400BadRequest'
    put:
      tags:
        - User
      summary: Update your user
      description: Update your user
      operationId: updateUser
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SimpleUser'
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleUser'
        400:
          $ref: '#/components/responses/400BadRequest'
    delete:
      tags:
        - User
      summary: Deletes a user
      description: delete a user
      operationId: deleteUser
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
      responses:
        204:
          description: 'Succesfull delete'
        400:
          $ref: '#/components/responses/400BadRequest'
  /updatePassword:
    put:
      tags:
        - User
      summary: Update a password
      description: Return your password
      operationId: updatePassword
      security: [ ]
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
        - name: token
          in: query
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  example: your_password
      responses:
        200:
          description: Successful operation
        400:
          $ref: '#/components/responses/400BadRequest'
    post:
      tags:
        - User
      summary: Request update a password
      description: Request update a password, it send an email and the user could change its password for an hour
      operationId: requestUpdatePassword
      security: [ ]
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
      responses:
        200:
          description: Successful operation
        400:
          $ref: '#/components/responses/400BadRequest'
...
  /users/{email}/plan/{plan}:
    put:
      tags:
        - User
      summary: Update your plan
      description: Update your plan
      operationId: updatePlan
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
        - name: plan
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SimpleUser'
        400:
          $ref: '#/components/responses/400BadRequest'
...
components:
  schemas:
...
    SimpleUser:
      type: object
      properties:
        userId:
          $ref: '#/components/schemas/Id'
        email:
          type: string
          example: theUser
        name:
          type: string
          example: John
        lastName:
          type: string
          example: James
        location:
          type: string
        description:
          type: string
        birthDay:
          type: string
          format: date
    UserOut:
      allOf:
        - $ref: '#/components/schemas/SimpleUser'
        - type: object
          properties:
            plan:
              type: integer
            guia:
              type: boolean
            age:
              type: integer
    UserCreate:
      allOf:
        - $ref: '#/components/schemas/SimpleUser'
        - type: object
          properties:
            password:
              type: string
              example: '12345'
    Id:
      type: integer
      format: int64
    Error:
      type: object
      properties:
        errorCode:
          type: string
        description:
          type: string
        cause:
          type: string
  responses:
    400BadRequest:
      description: Bad Request. Check the parameters or body of the request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    409Exists:
      description: Entity already exists with the same ID or combination of unique properties.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    basic:
      type: http
      scheme: basic
  headers:
    Link:
      description: >-
        The Link header provides information about the pagination of the response.
      schema:
        type: string
    X-Page-Number:
      description: >-
        The current page number of the response.
      schema:
        type: integer
    X-Total-Elements:
      description: >-
        The total number of elements available in the response.
      schema:
        type: integer
    X-Total-Pages:
      description: >-
        The total number of pages available in the response.
      schema:
        type: integer
security:
  - basic: [ ]